script/database_exists: silence expected errors
If the database doesn't exist, don't print an ugly error message.
That's the whole point of the command, after all: avoiding "database
[...] does not exist" errors by first checking if it exists.

If we get other kinds of errors, we'll still die correctly and show
those.

This was the behavior prior to 77ecc530b2f92ed90b5bd2b672b4235c4ca6da69.